requirements.txt
django==5.0.4
google-generativeai==0.4.0